Refactor lsp module into smaller modules for readability
