+++ src/extensions/clangd.rs
@@ -1,4 +1,5 @@
-use crate::{language_client::LanguageClient, utils::ToUrl};
+use crate::language_client::LanguageClient;
+use crate::utils::ToUrl;
 use anyhow::Result;
 use jsonrpc_core::Value;
 use lsp_types::{request::Request, TextDocumentIdentifier};
