+++ src/extensions/mod.rs
@@ -1,6 +1,7 @@
 pub mod clangd;
 pub mod gopls;
 pub mod java;
+pub mod rls;
 pub mod rust_analyzer;
 
 use crate::language_client::LanguageClient;
